OLLAMA-LANGCHAING-AGENTE/requirements.txt
# Core dependencies for the dynamic chat agent
langchain
langchain-core
langchain-ollama
langchain-deepseek
pyyaml
python-dotenv
